mantle/platform: make copied ISO file writable for tests
Since we are going to modify it let's ensure it's writable.

We are changing up some of the code that builds the Live artifacts
and it exposed the fact that we are relying here on the files in
the builddir to have the writable bit set. Let's not rely on that.

(cherry picked from commit 39cee5028f984b0e0f46f4fdbc0e778bdf61a82e)