build-deps.txt: add shellcheck
CI was updated[0] to use coreos-assembler as a buildroot instead of the fcos-buildroot container and
the openshift/os `validate` job[1] fails because shellcheck isn't installed in coreos-assembler. It
cannot be installed using `sudo` when running on OpenShift so it needs to be installed directly in
coreos-assembler.

[0] https://github.com/openshift/release/pull/59000/commits/68a7999da3e4fc98603b33b95cb83917b35e62d5
[1] https://github.com/openshift/os/blob/ce4da959eb3fc9477a0a96b69e6efd084a69b9ac/ci/prow-entrypoint.sh#L162

(cherry picked from commit 9090e0ffe8919e6c2e0676818409760eed821eb5)